fetch error, conditonnal layout, loader'

+++ utils/sendToGrist.js
@@ -3,17 +3,19 @@ import {useForm} from '../store/useFormStore'
 
 export default async function sendToGrist (Form) {
     const deleteDoc = useForm.getState().deleteDocumentByUri
-    let hasError = false
+    const currentCounter = useForm.getState().counter
+    const addData =  useForm.getState().addData
     const apiKey = process.env.EXPO_PUBLIC_GRIST_API_KEY
     const docId = process.env.EXPO_PUBLIC_GRIST_DOC_ID
     const hostName = process.env.EXPO_PUBLIC_GRIST_HOST
 
-    throw new Error()
-        
-    for (const doc of Form.documents) {
 
+    
+    
+    await new Promise(resolve => setTimeout(resolve, 3000));
+    for (const [index, doc] of Form.documents.entries()) {
+        
         try {
-            
             const formData = new FormData()
             doc.uri.forEach((d, i) => {
                 formData.append('upload', {
@@ -53,16 +55,23 @@ export default async function sendToGrist (Form) {
                     ]
                 })
             })
+
+            if(index === 1) {
+                throw new Error((doc.type ).toString())  
+            }
+
+            if(!response.ok) {
+                throw new Error((doc.type ).toString())
+            }
             const result = await response.json();
             console.log("Succès Grist:", result);
             deleteDoc(doc.uri)
+            addData({counter: currentCounter + 1 })
+            
+            
 
         } catch(error) {
-            hasError = true
-        }
-
-        if (hasError) {
-            throw new Error()
+            throw new Error(`Erreur lors de l'envoi du document "${error.message}".` || "Erreur lors de l'envoi des documents.")
         }
     }
